+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.1.1-beta.1] - 2020-04-03
+### Changed
+- Copy to clipboard after converting on the demo website.
+
+### Fixed
+- Double quotes instead of single quotes in the Docker CLI commands so we can reference environment variables.
+- Ordering of deletion of `docker network`.
 
 ## [0.1.0-beta.4] - 2020-04-02
 ### Added
@@ -43,7 +50,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial Release.
 - Convert docker-compose to Docker CLI syntax.
 
-[Unreleased]: https://gitlab.com/hmajid2301/composerisation/-/compare/release%2F0.1.0-beta.4...master
+[Unreleased]: https://gitlab.com/hmajid2301/composerisation/-/compare/release%2F0.1.1-beta.1...master
+[0.1.1-beta.1]: https://gitlab.com/hmajid2301/composerisation/-/tags/release%2F0.1.1-beta.1...release%2F0.1.0-beta.4
 [0.1.0-beta.4]: https://gitlab.com/hmajid2301/composerisation/-/tags/release%2F0.1.0-beta.3...release%2F0.1.0-beta.2
 [0.1.0-beta.3]: https://gitlab.com/hmajid2301/composerisation/-/tags/release%2F0.1.0-beta.3...release%2F0.1.0-beta.2
 [0.1.0-beta.2]: https://gitlab.com/hmajid2301/composerisation/-/tags/release%2F0.1.0-beta.2...release%2F0.1.0-beta.1